+++ Sources/ParseSwift/Types/Query.swift
@@ -123,6 +123,26 @@ public func == <T>(key: String, value: T) -> QueryConstraint where T: Encodable 
 }
 
 /**
+ Add a constraint that requires that a key is equal to a `ParseObject`.
+ - parameter key: The key that the value is stored in.
+ - parameter value: The `ParseObject` to compare.
+ - returns: The same instance of `QueryConstraint` as the receiver.
+ */
+public func == <T>(key: String, value: T) throws -> QueryConstraint where T: ParseObject {
+    let constraint: QueryConstraint!
+    do {
+        constraint = try QueryConstraint(key: key, value: value.toPointer())
+    } catch {
+        guard let parseError = error as? ParseError else {
+            throw ParseError(code: .unknownError,
+                             message: error.localizedDescription)
+        }
+        throw parseError
+    }
+    return constraint
+}
+
+/**
  Add a constraint that requires that a key is not equal to a value.
  - parameter key: The key that the value is stored in.
  - parameter value: The value to compare.
@@ -1630,6 +1650,17 @@ enum RawCodingKey: CodingKey {
     }
     init?(intValue: Int) {
         fatalError()
+    }
+}
+
+// MARK: CustomDebugStringConvertible
+extension Query {
+    public var debugDescription: String {
+        guard let descriptionData = try? ParseCoding.jsonEncoder().encode(self),
+            let descriptionString = String(data: descriptionData, encoding: .utf8) else {
+                return "\(className)"
+        }
+        return "\(className) (\(descriptionString))"
     }
 }
 // swiftlint:disable:this file_length