+++ Tests/ParseSwiftTests/ParseCloudTests.swift
@@ -99,6 +99,14 @@ class ParseCloudTests: XCTestCase { // swiftlint:disable:this type_body_length
         XCTAssertEqual(decoded, expected, "\"functionJobName\" key should be skipped by ParseEncoder")
     }
 
+    #if !os(Linux) && !os(Android)
+    func testDebugString() {
+        let cloud = Cloud2(functionJobName: "test", customKey: "parse")
+        let expected = "{\"customKey\":\"parse\",\"functionJobName\":\"test\"}"
+        XCTAssertEqual(cloud.debugDescription, expected)
+    }
+    #endif
+
     func testCallFunctionCommand() throws {
         let cloud = Cloud(functionJobName: "test")
         let command = cloud.runFunctionCommand()