+++ Tests/ParseSwiftTests/ParseQueryTests.swift
@@ -255,6 +255,12 @@ class ParseQueryTests: XCTestCase { // swiftlint:disable:this type_body_length
         XCTAssertEqual(decoded, expected)
     }
 
+    func testQueryEncoding() throws {
+        let query = GameScore.query()
+        let expected = "GameScore ({\"limit\":100,\"skip\":0,\"_method\":\"GET\",\"where\":{}})"
+        XCTAssertEqual(query.debugDescription, expected)
+    }
+
     func testFindExplainCommand() throws {
         let query = GameScore.query()
         let command: API.NonParseBodyCommand<Query<ParseQueryTests.GameScore>,
@@ -935,6 +941,22 @@ class ParseQueryTests: XCTestCase { // swiftlint:disable:this type_body_length
             return
         }
     }
+
+    func testWhereKeyEqualToParseObjectError() throws {
+        let compareObject = GameScore(score: 11)
+        XCTAssertThrowsError(try GameScore.query("yolo" == compareObject))
+    }
+
+    #if !os(Linux) && !os(Android)
+    func testWhereKeyEqualToParseObject() throws {
+        var compareObject = GameScore(score: 11)
+        compareObject.objectId = "hello"
+        let query = try GameScore.query("yolo" == compareObject)
+        // swiftlint:disable:next line_length
+        let expected = "GameScore ({\"limit\":100,\"skip\":0,\"_method\":\"GET\",\"where\":{\"yolo\":{\"__type\":\"Pointer\",\"className\":\"GameScore\",\"objectId\":\"hello\"}}})"
+        XCTAssertEqual(query.debugDescription, expected)
+    }
+    #endif
 
     func testWhereKeyNotEqualTo() {
         let expected: [String: AnyCodable] = [